// restat/restat-frontend/src/services/types/general.ts
export enum GroupBy {
  Profile = 'Group By Profile',
  Bidder = 'Group By Business Developer',
}

export enum DATE_OPTIONS {
  TODAY = 'Today',
  YESTERDAY = 'Yesterday',
  THIS_WEEK = 'This Week',
  LAST_WEEK = 'Last Week',
  THIS_MONTH = 'This Month',
  LAST_MONTH = 'Last Month',
  THIS_YEAR = 'This Year',
  LAST_YEAR = 'Last Year',
  ALL_DATA = 'All Data',
}